+++ rtp/JitterEstimator.h
@@ -7,48 +7,6 @@
 namespace rtp
 {
 
-template <typename T, size_t S>
-class Backlog
-{
-public:
-    T add(T value)
-    {
-        _index = (_index + 1) % S;
-        auto prev = _values[_index];
-        _values[_index] = value;
-        return prev;
-    }
-
-    T front() { return _values[_index]; }
-    T back() { return _values[(_index + 1) % S]; }
-
-    T getMean() const
-    {
-        T acc = 0;
-        for (int i = 0; i < S; ++i)
-        {
-            acc += _values[i];
-        }
-
-        return acc / S;
-    }
-
-    T getVariance(T hypotheticalMean) const
-    {
-        T acc = 0;
-        for (size_t i = 0; i < S; ++i)
-        {
-            auto d = _values[i] - hypotheticalMean;
-            acc += d * d;
-        }
-
-        return acc / S;
-    }
-
-private:
-    T _values[S];
-    uint32_t _index;
-};
 /**
  * Estimates the latency from jitter. That is the jitter buffer level you will need to be able to replay
  * media without having gaps due to jitter. The 95percentile level will calculate level needed to replay >95%
